FT: Add a getUIDs() method to requestLogger
The request Logger was previously managing his UIDs internally, without
providing any way for the user to retrieve the UIDs and send them to another
component, as metadata about the Request (thus allowing tracking a sub-request,
by namespacing it with the calling request's UIDs).

This is now provided, along with a set of tests that ensure more details about
the behavior of the UID list, as well as ensure that it's immutable through a
call to the getUids() method.