dev(config): Use an options object for config
In order to make the streams configurable for werelogs, it is preferable to
have a real options objects, documented in the JSDoc comment, rather than have
a never-ending list of parameters for the Logger's constructor.

As such, the log level, dump Threshold, and streams are now configurable
trhough the Logger's constructor's second parameter which is now an options
object. The streams shall be an array of streams into which bunyan can log
stuff. Their individual log levels are not configurable, as werelogs is
managing that himself on a global level.