dev(ModuleLogger): Add Module-level logging capabilities
Provide a module-level logging capability to werelogs, so that an user can log
through werelogs for module-level logging, instead of erroneously using a
requestLogger that will keep things in memory, and have an inconsistent
behavior with the wished-for situation.

Add associated tests for the module-logger, following the RequestLogger's tests
model.

Fixes #18