dev(RequestLogger): Add the posibility to set default JSON attributes
This commit provides the user of the RequestLogger with the capacity to
set JSON attributes to be included in all next log entries by default.
Currently, we do not plan on allowing to remove any of those, as the
use of this function is designed for request-long relevant information.

Any other kind of information should not be added as a default attribute to the
request logger.

 - Added the addDefaultFields method to RequestLogger class
 - Updated log entry generation to include new defined default fields
 - Updated README to show a simple example of using addDefaultFields
 - Added some tests to ensure that the fields are properly added