dev(RequestLogger): Add API to provide elapsed-time(ms) automatically
This commit provides a new API/Feature for the RequestLogger. Whenever a new
request logger is instanciated, an internal startTime(hrtime) is then retrieved
from the system, and stored in the instance. When the request reaches the end
of its processing, the user can then call upon the 'end' logging method, that
will automatically compute the elapsed time in ms, and include it in the
additional data fields. The logging level of this 'end' call can be configured
through the options Object passed to the ModuleLogger for configuration, but
the default of 'info' should be a good fit for most applications.

This commit provides:
 - 'end' config parameters to configure which log level to use for the 'end'
 call.
 - end logging function in the RequestLogger to provide a simple way to
 automatically include the hrtime diff into the data fields of the logging
 entry.
 - A test that checks the behavior of the RequestLogger.end() method.

Fixes #37