dev(RequestLogger): Rework end methods
This commit provides a new way to log the last entry of a RequestLogger.

The previous way was to configure a log level for the end method, as well as
providing an endError method to explicitly use the error level for this last
entry.

The new way is that the end method returns a RequestLogger-like object, from
which all the usual logging methods can be used as the user wishes (trace,
debug, info, warn, error, fatal). This should be easier to use, as the wrapped
object can be transmitted to other pieces of the code, and then used as any
other logger, with the only limitation that it can only be used once, since
it's dedicated to the _last_ entry written into a logger.

As such here is a list of things provided by this commit:
 - Create a new EndLogger that wraps around the RequestLogger for this specific
 last log entry:
   * end() now returns an internal wrapper object designed to augment last
   logging request.
   * end()'s wrapper provides access to all logging levels, for the user to use
   whichever he wants with the usual limitation of only one call (since it
   should be the last logging request)
   * end()'s wrapper provides an addDefaultFields() method to augment the data
   fields set only for the last logging entry.
 - Deprecate RequestLogger.end() and RequestLogger.endError() as they were
 defined before
 - Make new and old versions coexist in the code, until the next major release