FT: add time field
Normally node-bunyan automatically adds this field if it's omitted,
but it uses lodash to generate the timestamp, which in turn uses
a costly regex to figure out if the date method is available in an
env. That particular call was consuming a lot of cpu ticks, sitting
up top on the profiler log.