Provide a new Werelogs API
The former API presented a defect that required updating the design.
Indeed, due to the update() behavior of the instanciation of any new
werelogs Logger, a dependency could potentially override the original
configuration, making logs hard to control.

This design offers a new layer on top of the previous one, for the
following benefits:
 - Backwards compatibility on the handling of the Configuration:
   * Scope of the configuration update is a whole "API"
     (encompassing Loggers spawned from it, and their children
     objects)
   * Backwards compatibility on the way to quickly instanciate
     a Logger object
 - Configuration is no longer global, avoiding the raised issue
 - Provide a convenience API for the users that may not want to
   handle the API instance directly (usually not necessary)