impr: WRLGS-8 replace asserts by error logs
When RequestLogger.end() was called more than once for a particular
RequestLogger, an assert used to be raised, now an error log is dumped
instead which doesn't raise an exception and allows the worker to
continue processing further requests.

In this case an assert is not warranted because:

- the issue is not critical and doesn't prevent the process from
  continuing normal processing

- the issue could be triggered from unfrequent code paths that are
  hard to find by just reading the code.

By logging an error instead, we have a notification of the issue in
the process logs, and it helps to spot the root cause when
investigating because the 'error' log triggers a dump of the full
request trace logs.